Removed `get_optimal_mapping` (The solved is used in `GED` directly instead
