+++ pyged/ged.py
@@ -8,7 +8,7 @@ from typing import Optional, Tuple, Dict, Any
 import numpy as np
 import networkx as nx
 from pyged.costfunctions import CostFunction, ConstantCostFunction
-from pyged.bipartiteGED import compute_bipartite_cost_matrix, get_optimal_mapping, convert_mapping
+from pyged.bipartiteGED import compute_bipartite_cost_matrix, convert_mapping
 from pyged.solvers import Solver, SolverLSAP
 
 
@@ -81,7 +81,7 @@ class GED():
         # TODO : à sortir
         if ((rho is None) or (varrho is None)):
             C = compute_bipartite_cost_matrix(G1, G2, self.cf)
-            r, v = get_optimal_mapping(C, lsap_solver=self.solver)
+            r, v = self.solver.solve(C)
             rho, varrho = convert_mapping(r, v, G1, G2)
 
         ccf = self.cf if isinstance(self.cf, ConstantCostFunction) else self.cf.ccf
